refactor: Update UI layout, script asset paths, route middleware syntax, and product price column type.

+++ database/migrations/2024_01_08_024350_create_produks_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama_produk');
             $table->text('deskripsi');
             $table->string('foto');
-            $table->decimal('harga', 8, 2);
+            $table->integer('harga');
             $table->timestamps();
         });
     }
